+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Homepage - Fix air quality direction in RtL languages
 - Homepage - Fix missing hourly visibility in some cases
 - Settings - Material 3 Expressive theme
+- Settings - Add shortcuts to daily/trend configuration from cards configuration
 - Fix tint of “Open in another app” icon in landscape mode
 - Improve the formatting of today/tomorrow notification